增加范围支持

+++ MineKiller/Killer.cs
@@ -3,12 +3,16 @@ using StardewModdingAPI;
 using StardewValley;
 using StardewValley.Locations;
 using StardewValley.Monsters;
+using System.Threading;
+using Microsoft.Xna.Framework;
+using StardewValley.Characters;
+using System;
+using System.Threading.Tasks;
 
 namespace MineKiller
 {
     internal class Killer
     {
-
         private IModHelper _helper;
 
         private IMonitor _monitor;
@@ -16,7 +20,7 @@ namespace MineKiller
         private Config _config;
 
 
-        public void InitKiller(IModHelper helper,IMonitor monitor,Config config)
+        public void InitKiller(IModHelper helper, IMonitor monitor, Config config)
         {
             _helper = helper;
             _monitor = monitor;
@@ -34,11 +38,12 @@ namespace MineKiller
             {
                 return;
             }
+
             if (location is not MineShaft && location is not VolcanoDungeon && location is not Woods)
             {
                 return;
             }
-            
+
             NetCollection<NPC> characters = location.characters;
             for (var i = 0; i < characters.Count; i++)
             {
@@ -47,22 +52,19 @@ namespace MineKiller
                 {
                     continue;
                 }
+
                 Monster monster = character as Monster;
-                MakeWeek(monster,player);
-                _monitor.Log($"gonna kill monster [{monster.Name}]",LogLevel.Debug);
-                location.damageMonster(monster.GetBoundingBox(), _config.Damage, _config.Damage * 2, true, player);
+                if (IsIncludeRange(player.Position, monster.Position, _config.Range))
+                {
+                    MakeWeak(monster);
+                    _monitor.Log($"gonna kill monster [{monster.Name}]", LogLevel.Debug);
+                    location.damageMonster(monster.GetBoundingBox(), _config.Damage, _config.Damage * 2, monster is Mummy, player);
+                }
             }
         }
 
-        private void MakeWeek(Monster monster,Farmer player)
+        private void MakeWeak(Monster monster)
         {
-
-            Tool currentTool = null;
-
-            if (player.CurrentTool != null && player.CurrentTool is Tool)
-                currentTool = Game1.player.CurrentTool;
-
-
             if (monster is Bug bug && bug.isArmoredBug.Get())
             {
                 _helper.Reflection.GetField<NetBool>(bug, "isArmoredBug").SetValue(new NetBool(false));
@@ -73,8 +75,23 @@ namespace MineKiller
                 _helper.Reflection.GetField<NetBool>(rockCrab, "shellGone").SetValue(new NetBool(true));
                 _helper.Reflection.GetField<NetInt>(rockCrab, "shellHealth").SetValue(new NetInt(0));
             }
-
         }
 
+
+        public double CalculateLineDistance(Vector2 a, Vector2 b)
+        {
+            double line = 0;
+            //√（x₁-x₂）²+（y₁-y₂）²
+            double x = Math.Pow(a.X - b.X, 2);
+            double y = Math.Pow(a.Y - b.Y, 2);
+            line = Math.Sqrt(x + y);
+            return line;
+        }
+
+        private bool IsIncludeRange(Vector2 a, Vector2 b, int range)
+        {
+            double distance = CalculateLineDistance(a, b);
+            return distance < range;
+        }
     }
 }
